convert: playlist: Move M3UFile class to separate
module in util package.
